+++ 1.12/src/main/java/panszelescik/coolgadgets/gui/GuiHandler.java
@@ -1,6 +1,7 @@
 package panszelescik.coolgadgets.gui;
 
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.util.EnumHand;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.network.IGuiHandler;
 
@@ -18,7 +19,7 @@ public class GuiHandler implements IGuiHandler {
 	@Override
 	public Object getClientGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
 		if (ID == CALCULATOR)
-			return new CalculatorGUI();
+			return new CalculatorGUI(player.getHeldItem(EnumHand.MAIN_HAND));
 		return null;
 	}
 }