+++ 1.12/src/main/java/panszelescik/coolgadgets/items/ItemCalculator.java
@@ -13,6 +13,7 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
@@ -20,6 +21,7 @@ import panszelescik.coolgadgets.CoolGadgets;
 import panszelescik.coolgadgets.gui.GuiHandler;
 import panszelescik.morelibs.api.Helper;
 import panszelescik.morelibs.api.ItemBase;
+import panszelescik.morelibs.api.NBTHelper;
 
 public class ItemCalculator extends ItemBase {
 
@@ -28,7 +30,6 @@ public class ItemCalculator extends ItemBase {
 		setTranslationKey(MODID + ".calculator");
 		setRegistryName(new ResourceLocation(MODID, "calculator"));
 		setMaxStackSize(1);
-		setContainerItem(this);
 	}
 	
 	@Override
@@ -41,5 +42,8 @@ public class ItemCalculator extends ItemBase {
 	@SideOnly(Side.CLIENT)
 	public void addInformation(ItemStack stack, @Nullable World worldIn, List<String> tooltip, ITooltipFlag flagIn) {
 		tooltip.add(Helper.translate(getTranslationKey() + ".tooltip"));
+		if (NBTHelper.keyExists(stack, "display") && NBTHelper.keyExists(stack, "firstTyped")) {
+			tooltip.add(TextFormatting.ITALIC + Helper.translate(getTranslationKey() + ".tooltip2"));
+		}
 	}
 }