+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Cool Gadgets Changelog
 
+### 1.12.2-1.1.0
+- Requires More Libs 1.0.4+
+- Added Calculator's screen saving to NBT
+
 ### 1.12.2-1.0.5
 - Requires More Libs 1.0.3+
 - Actually Additions support, to link Lasers you must sneaking!